src/camera_ulilities/requirements.txt
numpy==1.24.2
numpysocket==1.0.1
PyYAML==5.3.1
opencv-python==4.7.0.72